refactor: seperate select merchant components
